Jackson causing big problems

+++ src/main/java/dist2/rest/bankclient/AccountRepo.java
@@ -1,5 +1,5 @@
 package dist2.rest.bankclient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface AccountRepo extends JpaRepository<Account, Integer> {
+interface AccountRepo extends JpaRepository<Account, Long> {
 }
